test(utxo-lib): move key utils to `test/testutil`
Issue: BG-38773
