+++ modules/utxo-lib/test/testutil/index.ts
@@ -1,0 +1,1 @@
+export * from './keys';
